+++ docs/changelog.rst
@@ -7,6 +7,7 @@ v1.3.0 (in development)
 -----------------------
 - Support Python 3.12
 - Update for PEP 714
+- Gave `PyPISimple` a `~PyPISimple.get_package_metadata()` method
 
 
 v1.2.0 (2023-09-23)